add folder for shared functions and standardise timeouts
