+++ .github/scripts/create-artifacts.sh
@@ -8,6 +8,7 @@ echo "Creating TCFV2 artifact..."
 mkdir -p tcfv2/nodejs/node_modules
 cd tcfv2
 cp ../../src/cmp_tcfv2.js nodejs/node_modules/pageLoadBlueprint.js
+cp -r ../../src/utils nodejs/node_modules
 zip -r nodejs.zip nodejs
 rm -rf nodejs
 cd ..
@@ -17,6 +18,7 @@ echo "Creating CCPA artifact..."
 mkdir -p ccpa/nodejs/node_modules
 cd ccpa
 cp ../../src/cmp_ccpa.js nodejs/node_modules/pageLoadBlueprint.js
+cp -r ../../src/utils nodejs/node_modules
 zip -r nodejs.zip nodejs
 rm -rf nodejs
 cd ..
@@ -26,6 +28,7 @@ echo "Creating AUS artifact..."
 mkdir -p aus/nodejs/node_modules
 cd aus
 cp ../../src/cmp_aus.js nodejs/node_modules/pageLoadBlueprint.js
+cp -r ../../src/utils nodejs/node_modules
 zip -r nodejs.zip nodejs
 rm -rf nodejs
 cd ..
