+++ src/utils/constants.js
@@ -1,0 +1,14 @@
+/** Ten (10) seconds in milliseconds */
+const TEN_SECONDS = 10000;
+
+/** Two (2) seconds in milliseconds */
+const TWO_SECONDS = 2000;
+
+/** Twenty (20) seconds in milliseconds */
+const TWENTY_SECONDS = 20000;
+
+module.exports = {
+	TEN_SECONDS,
+	TWO_SECONDS,
+	TWENTY_SECONDS,
+};
